Find way to make all the empires in the galaxy colonize without tanking performance. Delete FillPlanetDataForTechLevel because it's unnecesary and create a cache for GetCreatures to improve performance. Change how speedConfig works.

+++ src/Ai Empire Colonization/cEmpireTerraformingManager.h
@@ -109,7 +109,7 @@ private:
 
 	int elapsedTime;
 
-	int terraformingPerHour;
+	float terraformingPerHour;
 
 	bool terraformAllowed;
